minor cleanup, folder restructure, added material
