+++ src/app/helpers/typeof.pipe.ts
@@ -1,0 +1,8 @@
+import { Pipe, PipeTransform } from '@angular/core';
+
+@Pipe({
+  name: 'typeof'
+})
+export class TypeofPipe implements PipeTransform {
+  transform(value) { return typeof value; }
+}
