--- src/app/typings.ts
@@ -1,12 +1,0 @@
-export const DATA = {
-  Response: {
-    FilteredKeys: [
-      'main',
-      'weather',
-      'wind'
-    ],
-    SanitizedKeys: {
-      Weather: 'weather'
-    }
-  }
-};
