+++ src/app/typings/static.data.ts
@@ -1,0 +1,15 @@
+export const DATA = {
+  Request: {
+    Endpoint: 'https://api.openweathermap.org/data/2.5/weather'
+  },
+  Response: {
+    FilteredKeys: [
+      'main',
+      'weather',
+      'wind'
+    ],
+    SanitizedKeys: {
+      Weather: 'weather'
+    }
+  }
+};
